#19 [feat] 메모 상태관리 구현
