+++ src/stores/config/configStore.ts
@@ -3,6 +3,7 @@ import { configureStore } from '@reduxjs/toolkit';
 import modalToggle from '../modules/todoSlice';
 import counterSlice from '../modules/counterSlice';
 import categorySlice from '@stores/modules/category';
+import memoSlice from '@stores/modules/memo';
 
 //1. configureStore 함수를 사용하여 store를 생성
 //2. configureStore 함수의 인자로는 reducer 속성이 포함된 객체(module)가 전달
@@ -11,6 +12,7 @@ const store = configureStore({
     modalToggle,
     counterSlice,
     categorySlice,
+    memoSlice,
   },
 });
